requirements.txt
sentence-transformers
numpy
scikit-learn
ollama
faiss-cpu